Updates for Python 3

+++ image_misc.py
@@ -158,11 +158,11 @@ def crop_to_square(frame):
     i_size,j_size = frame.shape[0],frame.shape[1]
     if j_size > i_size:
         # landscape
-        offset = (j_size - i_size) / 2
+        offset = int((j_size - i_size) / 2)
         return frame[:,offset:offset+i_size,:]
     else:
         # portrait
-        offset = (i_size - j_size) / 2
+        offset = int((i_size - j_size) / 2)
         return frame[offset:offset+j_size,:,:]
 
 
@@ -575,7 +575,7 @@ def cv2_typeset_text(data, lines, loc, between = ' ', string_spacing = 0, line_s
                 if fs.align == 'right':
                     locx += fs.width - boxsize[0]
                 elif fs.align == 'center':
-                    locx += (fs.width - boxsize[0])/2
+                    locx += int((fs.width - boxsize[0])/2)
             #print 'right boundary is', locx + boxsize[0], '(%s)' % fs.string
                     #                print 'HERE'
             right_edge = locx + boxsize[0]
@@ -600,7 +600,7 @@ def cv2_typeset_text(data, lines, loc, between = ' ', string_spacing = 0, line_s
                 elif fs.align == 'left':
                     locx += fs.width
                 elif fs.align == 'center':
-                    locx += fs.width - (fs.width - boxsize[0])/2
+                    locx += int(fs.width - (fs.width - boxsize[0])/2)
             else:
                 locx += boxsize[0]
             locx += string_spacing
